Jira/Confluence: Use upstream database + license logic
Before the official Atlassian Helm charts were implemented, we needed to
implement an own logic for configuring the Jira and Confluence
instances. Now, with new features in the official Docker images and Helm
charts, we can drop some of the functionality again.

These features are not available for Crowd, thus Crowd is not affected.